ARM: dts: sun8i: a83t: Sync VBUS power supplies from Linux
These boards use CONFIG_USB0_VBUS_DET="AXP0-VBUS-DETECT" in their
defconfigs, but were missing the equivalent device tree property.
The device tree property is needed as the USB PHY driver will be
switching to use it instead of a virtual GPIO.

Copy the property (and the node it refers to) from the Linux v5.14
versions of the files.

Signed-off-by: Samuel Holland <samuel@sholland.org>